rename stores to store
