+++ src/store/boundStore.ts
