+++ src/store/imageSlice.ts
