+++ src/store/pointsSlice.ts
